requirements.txt
braintree==3.23.0
defusedxml==0.4.1
Django==1.8.7
django-allauth==0.24.1
django-braces==1.8.1
django-crispy-forms==1.5.2
django-filter==0.12.0
django-model-utils==2.3.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
httpie==0.9.3
jsonfield==1.0.3
Markdown==2.6.5
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
Pygments==2.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.1
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wheel==0.24.0
